rename result selectors to helper functions and move them out
[#INT_537]
