+++ app/javascript/selectors/index.js
@@ -6,3 +6,10 @@ export const findByCategory = (statusCodes = List(), selectedCategory) => (
 export const findByCode = (statusCodes, selectedCode) => (
   statusCodes.find(({code}) => code === selectedCode) || {}
 )
+
+export const buildSelector = (...funcs) => (
+  (arg) => {
+    var selector = funcs.pop()
+    return selector(...(funcs.map((f) => (f(arg)))))
+  }
+)
