+++ app/javascript/selectors/peopleSearchSelectors.js
@@ -1,12 +1,12 @@
-import {Map, List, fromJS} from 'immutable'
-import {findByCode} from 'selectors'
-
-const buildSelector = (...funcs) => (
-  (arg) => {
-    var selector = funcs.pop()
-    return selector(...(funcs.map((f) => (f(arg)))))
-  }
-)
+import {Map} from 'immutable'
+import {
+  mapLanguages,
+  mapIsSensitive,
+  mapIsSealed,
+  mapRaces,
+  mapEthnicities,
+  mapAddress,
+} from 'utils/peopleSearchHelper'
 
 const getPeopleSearchSelector = (state) => state.get('peopleSearch')
 export const getSearchTermValueSelector = (state) => (
@@ -19,71 +19,6 @@ export const getLastResultsSortValueSelector = (state) => {
   const lastResult = getPeopleSearchSelector(state).get('results').last()
   return lastResult.get('sort').toJS()
 }
-
-export const getResultLanguagesSelector = (state, result) => buildSelector(
-  (state) => state.get('languages'),
-  () => (result.get('languages') || List()),
-  (statusCodes, languages) => (
-    languages
-      .sort((item) => item.get('primary'))
-      .map((language) => (
-        findByCode(statusCodes.toJS(), language.get('id')).value)
-      )
-
-  )
-)(state)
-
-export const getIsSensitiveSelector = (result) => (result.get('sensitivity_indicator', '').toUpperCase() === 'S')
-export const getIsSealedSelector = (result) => (result.get('sensitivity_indicator', '').toUpperCase() === 'R')
-
-export const getResultRacesSelector = (state, result) => buildSelector(
-  (state) => state.get('ethnicityTypes'),
-  (state) => state.get('raceTypes'),
-  (state) => state.get('unableToDetermineCodes'),
-  () => (result.getIn(['race_ethnicity', 'race_codes']) || List()),
-  () => result.get('unable_to_determine_code'),
-  (ethnicityTypes, raceTypes, unableToDetermineCodes, races, unableToDetermineCode) => {
-    if (unableToDetermineCode) {
-      return List([findByCode(unableToDetermineCodes.toJS(), unableToDetermineCode).value])
-    } else {
-      return races
-        .map((race) => (Map({
-          race: findByCode(raceTypes.toJS(), race.get('id')).value,
-          race_detail: findByCode(ethnicityTypes.toJS(), race.get('id')).value,
-        })))
-    }
-  }
-)(state)
-
-export const getResultEthnicitiesSelector = (state, result) => buildSelector(
-  (state) => state.get('hispanicOriginCodes'),
-  () => (result.getIn(['race_ethnicity', 'hispanic_codes']) || List()),
-  () => (result.getIn(['race_ethnicity', 'hispanic_origin_code'])),
-  (hispanicOriginCodes, ethnicities, hispanicLatinoOriginCode) => fromJS({
-    hispanic_latino_origin: findByCode(hispanicOriginCodes.toJS(), hispanicLatinoOriginCode).value,
-    ethnicity_detail: ethnicities.map((ethnicity) => ethnicity.get('description')).toJS(),
-  })
-)(state)
-
-export const getResultAddressSelector = (result) => buildSelector(
-  (result) => (result.get('addresses') || List()).isEmpty(),
-  (result) => result.getIn(['addresses', 0, 'city']),
-  (result) => result.getIn(['addresses', 0, 'state_code']),
-  (result) => result.getIn(['addresses', 0, 'zip']),
-  (result) => result.getIn(['addresses', 0, 'street_number']),
-  (result) => result.getIn(['addresses', 0, 'street_name']),
-  (addressesEmpty, city, stateCode, zip, streetNumber, streetName) => {
-    if (addressesEmpty) { return null } else {
-      return Map({
-        city: city,
-        state: stateCode,
-        zip: zip,
-        type: '', // TODO: Implement as part of #INT-537
-        streetAddress: `${streetNumber} ${streetName}`,
-      })
-    }
-  }
-)(result)
 
 const formatSSN = (ssn) => ssn && ssn.replace(/(\d{3})(\d{2})(\d{4})/, '$1-$2-$3')
 export const getPeopleResultsSelector = (state) => getPeopleSearchSelector(state)
@@ -99,17 +34,17 @@ export const getPeopleResultsSelector = (state) => getPeopleSearchSelector(state
       nameSuffix: result.get('name_suffix'),
       legacyDescriptor: result.get('legacy_descriptor'),
       gender: result.get('gender'),
-      languages: getResultLanguagesSelector(state, result),
-      races: getResultRacesSelector(state, result),
-      ethnicity: getResultEthnicitiesSelector(state, result),
+      languages: mapLanguages(state, result),
+      races: mapRaces(state, result),
+      ethnicity: mapEthnicities(state, result),
       dateOfBirth: fullResult.getIn(['highlight', 'date_of_birth', 0], result.get('date_of_birth')),
       ssn: formatSSN(fullResult.getIn(['highlight', 'ssn', 0], result.get('ssn'))),
-      address: getResultAddressSelector(result),
+      address: mapAddress(result),
       phoneNumber: phoneNumber && Map({
         number: phoneNumber.get('number'),
         type: phoneNumber.get('type'),
       }),
-      isSensitive: getIsSensitiveSelector(result),
-      isSealed: getIsSealedSelector(result),
+      isSensitive: mapIsSensitive(result),
+      isSealed: mapIsSealed(result),
     })
   })
