+++ spec/javascripts/selectors/peopleSearchSelectorsSpec.js
@@ -3,12 +3,6 @@ import {fromJS, Map} from 'immutable'
 import {
   getPeopleResultsSelector,
   getLastResultsSortValueSelector,
-  getResultLanguagesSelector,
-  getResultRacesSelector,
-  getResultEthnicitiesSelector,
-  getIsSensitiveSelector,
-  getIsSealedSelector,
-  getResultAddressSelector,
 } from 'selectors/peopleSearchSelectors'
 
 describe('peopleSearchSelectors', () => {
@@ -59,170 +53,7 @@ describe('peopleSearchSelectors', () => {
     })
   })
 
-  describe('getResultLanguagesSelector', () => {
-    it('maps languages to lov values by id, sorting by primary', () => {
-      const result = fromJS({
-        languages: [
-          {id: '3', primary: true},
-          {id: '2', primary: false},
-          {id: '1', primary: true}],
-      })
-      const state = fromJS({languages: languageLovs})
-      const languageResult = getResultLanguagesSelector(state, result)
-      expect(languageResult).toEqualImmutable(
-        fromJS(['French', 'English', 'Italian'])
-      )
-    })
-  })
-
-  describe('getIsSensitiveSelector', () => {
-    it('returns true if the result is sensitive', () => {
-      const result = fromJS({
-        sensitivity_indicator: 'S',
-      })
-      const sensitiveResult = getIsSensitiveSelector(result)
-      expect(sensitiveResult).toEqual(true)
-    })
-
-    it('returns false if the result is not sensitive', () => {
-      const result = fromJS({
-        sensitivity_indicator: 'R',
-      })
-      const sensitiveResult = getIsSensitiveSelector(result)
-      expect(sensitiveResult).toEqual(false)
-    })
-  })
-
-  describe('getIsSealedSelector', () => {
-    it('returns true if the result is sealed', () => {
-      const result = fromJS({
-        sensitivity_indicator: 'R',
-      })
-      const sensitiveResult = getIsSealedSelector(result)
-      expect(sensitiveResult).toEqual(true)
-    })
-
-    it('returns false if the result is not sealed', () => {
-      const result = fromJS({
-        sensitivity_indicator: 'S',
-      })
-      const sensitiveResult = getIsSealedSelector(result)
-      expect(sensitiveResult).toEqual(false)
-    })
-  })
-
-  describe('getResultAddressSelector', () => {
-    it('returns the city, state, zip, empty type, and a joined street address', () => {
-      const result = fromJS({
-        addresses: [{
-          city: 'city',
-          state_code: 'state',
-          zip: 'zip',
-          type: 'blah',
-          street_number: '123',
-          street_name: 'C Street',
-
-        }],
-      })
-      const addressResult = getResultAddressSelector(result)
-      expect(addressResult).toEqualImmutable(fromJS({
-        city: 'city',
-        state: 'state',
-        zip: 'zip',
-        type: '',
-        streetAddress: '123 C Street',
-      }))
-    })
-  })
-
-  describe('getResultRacesSelector', () => {
-    it('maps races to lov values by id', () => {
-      const result = fromJS({
-        race_ethnicity: {
-          race_codes: [
-            {id: '3', description: 'Romanian'},
-            {id: '2', description: 'French'},
-            {id: '1', description: 'European'},
-          ],
-        },
-      })
-      const state = fromJS({ethnicityTypes: ethnicityTypeLovs, raceTypes: raceTypeLovs})
-      const racesResult = getResultRacesSelector(state, result)
-      expect(racesResult).toEqualImmutable(
-        fromJS([
-          {race: 'Race 3', race_detail: 'Romanian'},
-          {race: 'Race 2', race_detail: 'French'},
-          {race: 'Race 1', race_detail: 'European'},
-        ])
-      )
-    })
-
-    it('maps races to "Abandoned" if unableToDetermineCode is "A"', () => {
-      const result = fromJS({
-        unable_to_determine_code: 'A',
-      })
-      const state = fromJS({unableToDetermineCodes})
-
-      const racesResult = getResultRacesSelector(state, result)
-      expect(racesResult).toEqualImmutable(
-        fromJS(['Abandoned'])
-      )
-    })
-
-    it('maps races to "Unknown" if unableToDetermineCode is "I"', () => {
-      const result = fromJS({
-        unable_to_determine_code: 'I',
-      })
-      const state = fromJS({unableToDetermineCodes})
-
-      const racesResult = getResultRacesSelector(state, result)
-      expect(racesResult).toEqualImmutable(
-        fromJS(['Unknown'])
-      )
-    })
-
-    it('maps races to "Unknown" if unableToDetermineCode is "K"', () => {
-      const result = fromJS({
-        unable_to_determine_code: 'K',
-      })
-      const state = fromJS({unableToDetermineCodes})
-
-      const racesResult = getResultRacesSelector(state, result)
-      expect(racesResult).toEqualImmutable(
-        fromJS(['Unknown'])
-      )
-    })
-  })
-
-  describe('getResultEthnicitiesSelector', () => {
-    it('maps hispanic codes to lov values', () => {
-      const result = fromJS({
-        race_ethnicity: {
-          hispanic_origin_code: 'Y',
-          hispanic_codes: [
-            {id: '3', description: 'Romanian'},
-            {id: '2', description: 'French'},
-            {id: '1', description: 'European'},
-          ],
-        },
-      })
-      const state = fromJS({hispanicOriginCodes, ethnicityTypes: ethnicityTypeLovs, raceTypes: raceTypeLovs})
-      const racesResult = getResultEthnicitiesSelector(state, result)
-      expect(racesResult).toEqualImmutable(
-        fromJS({
-          hispanic_latino_origin: 'yes', ethnicity_detail: ['Romanian', 'French', 'European']}
-        )
-      )
-    })
-  })
-
   describe('getPeopleResultsSelector', () => {
-    const languageLovs = [
-      {code: '1', value: 'English'},
-      {code: '2', value: 'French'},
-      {code: '3', value: 'Italian'},
-    ]
-
     it('maps person search attributes to suggestion attributes', () => {
       const peopleSearch = {
         results: [{
